add function to add the contents of the global symbol table to the lifted output
For BinaryNinja these allows us to automatically generate all the
information that the patcher needs, instead of having to feed it an
additional header file. There's a parameter to 'to_c' that controls
whether this should be done or not, to avoid breaking other
users. Only tested with 'normal' globals, arrays and functions, which
is what we have in challenge 9