allow choosing a different format when comparing cfgs
I don't know how to display a pdf file within the binary ninja UI,
but for a png file I can create a simple html and they have a way
to display that.

this adds an extra command line argument to relational compare cfgs
called --format. for now I only let the user choose pdf and png,
but I could extend it to all the formats allowed by dot if
we wanted to. the default continues to be pdf so nothing should
need to change on all the existing users. I tested both explicitly
asking for pdf and png and also the default and everything worked