Add ASTDeserializer method to read back return sequences
This adds a method to ASTDeserializer which can build an
ASTReturnSequences object. We use this so Binary Ninja can
add the return sequences computed by CodeHawk to its
lifting output so patching can succeed.

Along the way I also fixed a minor typo in the `__str__` method that was
making mypy complain.