Fix a number of bugs in the callgraph generation code
The biggest issue that is fixed is that some nodes use the function name
and some use the function address and it gets pretty confusing which
one we should be using. this is a bit of a hack in that we just check
both, but it feels like always using the same (probably address?)
would be best.

This also fixes a bug where we were constraining the sinks by the source
and not the destination, and a typo in the status message.

Finally, this streamlines the code a bit so we don't do the same check
on every iteration of the subloop that can be done in the outer loop