switch cfg comparison's json path to use the unified json file
instead of the custom json function from jsonresultutil. note that some of the
functions in there are used by the non-relational case so those cannot be deleted.

Also, the dot/pdf path does not use the json output to do its job yet. I'll
leave that to be tackled later on.

As part of a rebase, this change also updates the json output to match
the changes we made in 1bb2992, by adding the ID attribute to the
block's json content, and using that ID for edges and the block-mapping.