remove unused block semantic comparison
we represent block differences using cfgblockmappingitem schema
augmented to include information about the instructions added/removed/changed.

this also removes a bunch of unused fields in JSONBlockComparison which
are now present in the containing object.

I updated the schema to not include the old fields, but i need to figure
out how to add the new fields since they're nested quite a bit