remove unused things from JSONFunctionComparison
we've simplified it so it contains the two CFGs and the block-mapping,
which in turn contains all the block-level comparison information that
used to be exposed at this level.

localvars comparison never got implemented, same with semantic comparison.

this also updates the schemas to remove the unused ones, and document
what's missing after these changes.

This also removes a number of unused imports