remove JSONCfgComparison and associated classes
JSONFunctionComparison has absorbed all the cfg comparison details
so we no longer need a separate class for that. Some of the associated
classes were part of a new API based on graph edits that we never
finished so those are getting nuked too.

This also updates the schemas and json checker