Prevent the creation of multiple stackvariables with the same name
Before this change, whenever a stack location is used as an lhs, we were
blindly creating a new variable for it with whatever type we decided at
that point. However, if a stack location is used as a lhs in multiple
places with different types, what ends up happening in the code is we do
multiple definitions of variables with the same name but different
types, which makes everyone mad:
  int stackvar_1404;
  ...
  char stackvar_1404;

This changes the ASTInterface code to check first if we already have
a variable with the requested name, and if so, we reuse it. There
are some additional hacks in place to:
- if the existing variable doesn't have a type, but the new request
does, assign the variable that type.
- if the types are different, then make a note that we should probably
do a cast? Although now that I think about it, it may be fine to not
do a cast and avoid the complexity. This part of the change required
teaching ASTTyp and its subclasses how to do comparison, because
currently, each stackvar has its own instance of the ASTType object
which always evualuate to being not equal.
- Don't allow stack variables to be created with type void. Instead we
assign it type iuchar.