switch functionality to avoid creating duplicate variables into add_symbol
this way not only stack variables get the benefit. Alongside that,
add_symbol also prevents creating variables with void type. The biggest
functionality change is that while add_symbol already had some of this
functionality, it would create a new ASTVarInfo instead of reusing the
existing one. This switches that, and adds a couple of setters to
ASTVarInfo to make it clear that they're not immutable objects.

I tested on the contec and grididps and things seem to be working well