Switch from using jar to using zip
That way users don't need to install java to run us. Jar is
just a zip file so the change is purely cosmetic.

The ocaml side may be more of a rabbit hole, which then may
extend to the other analyzers.

There are still some things using the .jar extension:
- `bchsummaries.jar` since that file is a copy of what we have in the ocaml analyzer
repo:
https://github.com/static-analysis-engineering/codehawk/tree/master/CodeHawk/CHB/bchsummaries
https://github.com/static-analysis-engineering/codehawk/blob/master/CodeHawk/CHB/bchsummaries/makejar.sh
- `*functions.jar` in the analysis folder: the ocaml analyzer reads in
  this file and expects that filename:
https://github.com/static-analysis-engineering/codehawk/blob/master/CodeHawk/CHB/bchlib/bCHPreFileIO.ml#L183-L191